added extension check

+++ app.py
@@ -19,32 +19,36 @@ def scan():
     # fileitem = form.getvalue('fileName')
     _fileitem = request.files.get('fileName')
 
-    _fullname = _fileitem.filename
+    # _fullname = _fileitem.filename
     name = _fileitem.filename.split('.')[0]
+    extension = _fileitem.filename.split('.')[1]
+    if extension == 'mp4':
+        if _fileitem:
+            _fileitem.save(f'videos/{name}.mp4')
+            message = 'The file was uploaded successfully'
+        else:
+            message = 'No file was uploaded'
+            output = 'Try Again!'
+            return render_template('index.html', result = '{}'.format(output))
 
-    if _fileitem:
-        _fileitem.save(f'videos/{name}.mp4')
-        message = 'The file was uploaded successfully'
+        
+        print(message)
+        print(_fileitem)
+        # print(_fullname)
+
+        #passing for preprocessing
+        output = dp.dataProcess(name)
+
+        #deleting directories
+        print('Deleting files...')
+        os.remove(f'videos/{name}.mp4')
+        shutil.rmtree(f'videos/{name}/')
+        print('Files deleted after detection!')
+
+        return render_template('index.html', result = '{}'.format(output), loading = '{}'.format(" "))
+        
     else:
-        message = 'No file was uploaded'
-        output = 'Try Again!'
-        return render_template('index.html', result = '{}'.format(output))
-
-    
-    print(message)
-    print(_fileitem)
-    print(_fullname)
-
-    #passing for preprocessing
-    output = dp.dataProcess(name)
-
-    #deleting directories
-    print('Deleting files...')
-    os.remove(f'videos/{name}.mp4')
-    shutil.rmtree(f'videos/{name}/')
-    print('Files deleted after detection!')
-
-    return render_template('index.html', result = '{}'.format(output), loading = '{}'.format(" "))
+        return render_template('index.html', result = '{}'.format('Error!'), loading = '{}'.format(" "))
 
 
 if __name__=="main":
